Added Battleship game skeleton
As of right now it is just a Visual Studio project. We will need to add
it to Unity, but I wasn't exactly sure where it all needed to go and how
to do it.

See Program.cs to see how to use the engine.
The engine itself can be found in BattleShipCompetition.cs.