Set up Battleship grid with init script
The grid is now individual cubes that can be modified indivually. And
the view of the boards is slightly different now.

We can change the view if you would like, it was just hard to make it
all work in the old way. And now there is room for the chatroom.